+++ src/main/java/io/bookwright/di/UiModule.java
@@ -3,7 +3,6 @@ package io.bookwright.di;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.microsoft.playwright.Page;
-import io.bookwright.config.Configs;
 import io.bookwright.config.MainConfig;
 import io.bookwright.junit.TestUser;
 import io.bookwright.junit.UserFixtureExtension;
@@ -12,22 +11,11 @@ import org.junit.jupiter.api.extension.ExtensionContext;
 
 public class UiModule extends AbstractModule {
 
-  private final ExtensionContext context;
-
-  public UiModule(ExtensionContext context) {
-    this.context = context;
-  }
-
-  @Override
-  protected void configure() {
-    bind(MainConfig.class).toInstance(Configs.main());
-  }
-
   @Provides
-  Page page() {
+  Page page(ExtensionContext context, MainConfig config) {
     TestUser user = UserFixtureExtension.find(context).orElse(null);
     return user == null
         ? BrowserManager.page()
-        : BrowserManager.page(user.session(), Configs.main().localBookingBaseUrl());
+        : BrowserManager.page(user.session(), config.localBookingBaseUrl());
   }
 }